requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cssbeautifier==1.14.11
cycler==0.12.1
diff-match-patch==20230430
Django==4.2.7
django-admin-rangefilter==0.13.0
django-browser-reload==1.12.1
django-environ==0.11.2
django-filter==23.4
django-geojson==4.1.0
django-import-export==4.0.7
django-storages==1.14.4
django-tailwind==3.6.0
django-unfold==0.40.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djlint==1.34.0
drf-yasg==1.21.7
EditorConfig==0.12.3

google-auth==2.31.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsbeautifier==1.14.11
json5==0.9.14
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2

openapi-codec==1.3.2
openpyxl==3.1.2
OWSLib==0.30.0
packaging==24.0
pandas==2.2.2
pathspec==0.11.2
pillow==10.3.0
plotly==5.20.0
ply==3.11
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyfcm==2.0.4
Pygments==2.17.2
PyJWT==2.8.0

PyYAML==6.0.1
regex==2023.10.3
remotior_sensus==0.3.5
reportlab==4.1.0
requests==2.31.0
rich==13.7.0
rsa==4.9
ruff==0.1.6
scipy==1.13.0
setuptools==69.2.0
shapely==2.0.3
simplejson==3.19.2
sip==6.8.3
six==1.16.0
slippers==0.6.2
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0

